# openapi.yml
...
openapi: 3.1.0
info:
  title: Postmill API
  version: 3.4.0
  description: |
    Comprehensive OpenAPI description of the **Postmill** backend (NestJS) as it exists in
    `postmill-ai/postmill-app`.

    Postmill schedules social-media and chat posts to 28+ channels, with a calendar, analytics, team
    management, a media library, and — in this fork — a pluggable multi-provider **AI layer**,
    **Analytics v2** persisted dashboards, **social comments** sync/moderation, admin-configurable
    **channel OAuth credentials**, and an OAuth2 authorization-code provider.

    ## Surfaces & authentication

    There are three distinct authentication surfaces:

    1. **Dashboard API** (the bulk of routes). Requires a Postmill user session JWT supplied either as
       the `auth` **cookie** or the `auth` **request header**. The active organization is selected
       with the `showorg` cookie/header. Many routes are further gated by RBAC policies
       (`@CheckPolicies`) or by `isSuperAdmin`.
    2. **Public integration API** (`/public/v1/*`). Requires an organization **API key** (or a
       `pos_`-prefixed OAuth access token) in the `Authorization` header. Intended for
       n8n/Zapier/CLI/3rd-party automation.
    3. **Unauthenticated routes** — registration/login, the OAuth2 authorization/token endpoints,
       Stripe & enterprise webhooks, the social-connect OAuth callback, health/monitor, and public
...

...
    > global route prefix — controller paths are mounted at the root.

...

...
servers:
...
security:
  - cookieAuth: []
  - authHeader: []
paths:
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      operationId: listCampaigns
      responses: { '200': { description: Campaigns } }
    post:
      tags: [Campaigns]
      summary: Create a campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                color: { type: string }
                description: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
      responses: { '200': { description: Created } }
  /campaigns/{id}:
    get:
      tags: [Campaigns]
      summary: Get a campaign
      operationId: getCampaign
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Campaign } }
    put:
      tags: [Campaigns]
      summary: Update a campaign
      operationId: updateCampaign
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
                description: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
                archived: { type: boolean }
      responses: { '200': { description: Updated } }
    delete:
      tags: [Campaigns]
      summary: Delete a campaign
      operationId: deleteCampaign
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Deleted, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessFlag' } } } } }

  # ===================================================================== PROVIDER CAPABILITIES (fork)
...
  /integrations/mentions:
    post:
      tags: [Integrations]
      summary: Resolve @-mentions for a channel
      operationId: mentions
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IntegrationFunctionDto' } } }
      responses: { '200': { description: Mention candidates, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Session JWT in the `auth` cookie (dashboard API).
    authHeader:
      type: apiKey
      in: header
      name: auth
      description: Session JWT in the `auth` header (alternative to the cookie).
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Organization **API key** or a `pos_`-prefixed OAuth access token, sent **raw** (no `Bearer`
        prefix) in the `Authorization` header. Used by `/public/v1/*`.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    SuccessFlag:
      type: object
      properties:
        success: { type: boolean, example: true }
...
    IntegrationFunctionDto:
      type: object
      required: [name, id]
      properties:
        name: { type: string, description: Provider method name }
        id: { type: string, description: Integration id }
        data: { description: Arbitrary payload passed to the provider method }
...